Refactor structure with rollout
