+++ robusta_krr/core/integrations/kubernetes/__init__.py
@@ -21,7 +21,7 @@ from robusta_krr.core.models.objects import K8sObjectData, PodData
 from robusta_krr.core.models.result import ResourceAllocations
 from robusta_krr.utils.configurable import Configurable
 
-from robusta_krr.core.integrations.rollout import RolloutAppsV1Api
+from .rollout import RolloutAppsV1Api
 
 
 class ClusterLoader(Configurable):
